libc - err.h, исправлен размер pid_t

+++ kernel/kairax/types.h
@@ -30,7 +30,7 @@ typedef unsigned long long size_t;
 typedef long long ssize_t;
 typedef size_t loff_t;
 
-typedef long long 			pid_t;
+typedef int 				pid_t;
 typedef unsigned int 		uid_t;
 typedef unsigned int 		gid_t;
 typedef unsigned long int 	ino_t;
